# VxDesign: Reduce memory usage when rendering ballots