--- a/apps/design/backend/src/worker/ballot_pdfs.ts
+++ b/apps/design/backend/src/worker/ballot_pdfs.ts
@@ -1,15 +1,14 @@
 import {
   convertPdfToGrayscale,
   BaseBallotProps,
-  RenderDocument,
   calibrationSheetTemplate,
   Renderer,
 } from '@votingworks/hmpb';
 import { HmpbBallotPaperSize } from '@votingworks/types';
 
-export async function renderBallotPdf(
-  props: BaseBallotProps,
-  document: RenderDocument
+export async function normalizeBallotColorModeForPrinting(
+  ballotPdf: Uint8Array,
+  props: BaseBallotProps
 ): Promise<Uint8Array> {
   /**
    * Specific to NH V4 ballots with tinted headers/footers.
@@ -17,12 +16,11 @@ export async function renderBallotPdf(
    */
   const needsColorPrint = 'colorTint' in props && !!props.colorTint;
 
-  const colorPdf = await document.renderToPdf();
   if (needsColorPrint) {
-    return colorPdf;
+    return ballotPdf;
   }
 
-  return await convertPdfToGrayscale(colorPdf);
+  return await convertPdfToGrayscale(ballotPdf);
 }
 
 export async function renderCalibrationSheetPdf(