--- a/libs/fixture-generators/src/generate-election-package/generate_election_package.test.ts
+++ b/libs/fixture-generators/src/generate-election-package/generate_election_package.test.ts
@@ -35,7 +35,7 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
-  await renderer.cleanup();
+  await renderer.close();
 });
 
 const testCases = [