--- a/libs/fixture-generators/src/generate-election-package/generate_election_package.ts
+++ b/libs/fixture-generators/src/generate-election-package/generate_election_package.ts
@@ -88,7 +88,7 @@ export async function generateElectionPackage(
   zip.file(ElectionPackageFileName.ELECTION, electionDefinition.electionData, {
     date: FIXTURES_FILE_DATE,
   });
-  await renderer.cleanup();
+  await renderer.close();
 
   zip.file(
     ElectionPackageFileName.SYSTEM_SETTINGS,