… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-1-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-10-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-11-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-12-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-13-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-14-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-15-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-16-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-17-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-18-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-19-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-2-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-20-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-21-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-22-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-23-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-24-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-25-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-26-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-27-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-28-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-29-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-3-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-30-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-31-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-32-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-4-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-5-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-6-snap.png
… apps/design/backend/src/__image_snapshots__/test-decks-test-ts-create-precinct-test-deck-for-a-precinct-with-multiple-ballot-styles-7-snap.png