## Overview

Task: https://github.com/votingworks/vxsuite/issues/4914

The main memory bottleneck in the current code is that we keep a browser page open for each rendered ballot until we can extract the grid layouts, create the election definition, hash it, and inject that hash in the QR code.

I addressed this by splitting the process into phases: 
1. Render each ballot, then extract the grid layout and the HTML content. Close the browser page when done.
2. Use the grid layouts from all ballots to create the election definition (and its ballot hash)
3. Go through each ballot again, creating a new browser page and loading the previously created HTML content and injecting the QR code with the ballot hash

I also noticed that even with this approach, there appeared to be some memory bloat from creating a new browser page for each ballot, even if I cleaned up the page after using it. To fix this, I enforced rendering the ballots serially so we can reuse a single browser page. Making that change seemed to stop the memory bloat.

Moving forward, we can probably tune this approach more to better balance memory and speed by using a pool of browser pages, but I wanted to start with a minimal change that solved the memory issue.

## Demo Video or Screenshot
Some basic benchmarking with the electionGeneral fixture on my dev machine:

**main**
- export election package + ballots: 1GB memory, 13.051s
- export test decks: 2.5 GB memory, 5.527s

**this branch**
- export election package + ballots: 100-200MB, 14.271s
- test decks: 100-200MB, 20.329s

## Testing Plan
- Manual testing and benchmarking
- Updated existing tests

## Checklist

- [ ] I have prefixed my PR title with "VxDesign: ", "VxPollBook: ", or "HWTA: " if my change is specific to one of those products.
- [ ] I have added [logging](https://github.com/votingworks/vxsuite/tree/main/libs/logging) where appropriate for any new user actions.
- [ ] I have added the "user_facing_change" label to this PR, if relevant, to automate an announcement in #machine-product-updates.